Multiple LLMs save TXT
Multiple LLMs answers were integrated into the app with new firebase architecure, once it is again changed, this needs to be a bit revised.
